+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Remove copied duplicate ids
+
 ## [v1.19.0] - 2022-10-21
 
 * Change title for `Problems` in `organic-chemistry`